chore(code): move old code to DeprecatedCode
