+++ packages/hedgehog-lab/src/components/DeprecatedCode/Footer/Footer.tsx
